# src/main/java/gift/academic/promo_it/doc/swagger.yaml
openapi: 3.0.3
info:
  title: Promo-IT Backend
  description: "Spring Boot API для Promo-IT: регистрация пользователей с уникальным ADMIN и JWT аутентификация."
  version: 1.0.0
  contact:
    name: Promo-IT Team
servers:
  - url: http://localhost:8080
    description: Development server
tags:
  - name: Authentication
    description: Операции регистрации и входа
  - name: OTP
    description: Работа с кодами OTP.
  - name: Users
    description: Управление пользователями
  - name: Operations  # Add this
    description: Операции
paths:
  /users/list:
    get:
      x-role-required: ADMIN
      tags:
        - Users
      summary: Получить список всех пользователей (кроме администраторов)
      description: Возвращает список пользователей с ролью 'ordinary'. Доступно только для ADMIN.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Список пользователей успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '403':
          description: Доступ запрещен
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /users/{id}:
    delete:
      x-role-required: ADMIN
      tags:
        - Users
      summary: Удалить пользователя по ID
      description: Удаляет пользователя из системы. Доступно только для ADMIN. Администратора удалить нельзя (бизнес-логика).
      parameters:
        - name: id
          in: path
          required: true
          description: ID пользователя для удаления
          schema:
            type: integer
            format: int64
            minimum: 1
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Пользователь успешно удален
        '400':
          description: "Ошибка валидации: число меньше 1 или больше допустимого максимума"
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Не авторизован
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Доступ запрещен (недостаточно прав)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Пользователь не найден
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /otp/settings:
    get:
      x-role-required: ADMIN
      tags:
        - OTP
      summary: Получить текущую конфигурацию OTP
      description: Возвращает параметры OTP (lifespan и количество символов). Доступно только ADMIN.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Текущая конфигурация OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpConfigResponse'
        '403':
          description: Доступ запрещен (требуется роль ADMIN)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Не авторизован
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    put:
      x-role-required: ADMIN
      tags:
        - OTP
      summary: Обновить конфигурацию OTP
      description: Обновляет lifespan и number_of_symbols. Доступно только ADMIN. Таблица хранит не более одной записи.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpConfigUpdateRequest'
      responses:
        '200':
          description: Обновленная конфигурация OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpConfigResponse'
        '400':
          description: Валидация ошибок
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /otp/generate:
    post:
      tags:
        - OTP
      summary: Сгенерировать OTP код
      description: Генерирует и отправляет код согласно текущим настройкам. Доступно только авторизованным пользователям.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpGenerateRequest'
      responses:
        '202':
          description: Запрос принят к обработке
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpGenerateResponse'
        '400':
          description: Ошибка валидации тела запроса
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Не авторизован
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '422':
          description: Код некуда отправить (емейл неверен и т.п.).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /otp/validate:
    post:
      tags:
        - OTP
      summary: Проверить OTP код
      description: |
        Проверяет OTP код по ID операции. 
        Логика:
        1. Поиск активной записи по operationId.
        2. Проверка соответствия кода.
        3. Проверка срока действия (TTL).
        При успехе статус кода меняется на USED.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpValidateRequest'
      responses:
        '200':
          description: Код успешно проверен
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "OTP validated successfully"
        '400':
          description: Неверный формат данных или код не совпадает
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Код просрочен (EXPIRED) или не найден
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Не авторизован
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  /auth/register:
    post:
      tags:
        - Authentication
      summary: Регистрация пользователя
      description: Создаёт пользователя с ролью ordinary или admin (если admin ещё нет).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              validOrdinary:
                summary: Валидный обычный пользователь
                value:
                  login: "user"
                  password: "secretUserPass"
                  role: "ordinary"
                  email: "ordinary@example.com"
              validAdmin:
                summary: Создание админа (если ещё не существует)
                value:
                  login: "admin"
                  password: "secretAdminPass"
                  role: "admin"
                  email: "admin@examle.com"
      responses:
        '201':
          description: Пользователь успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Ошибка валидации (слабый пароль, логин занят, неполный набор данных и т.п.)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '409':
          description: Конфликт (роль Admin уже занята или логин уже существует)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Аутентификация пользователя
      description: Проверяет учетные данные и возвращает JWT-токен.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              ordinaryUser:
                summary: Обычный пользователь
                description: Аутентификация для пользователя с ролью ordinary
                value:
                  login: "user"
                  password: "secretUserPass"
              adminUser:
                summary: Администратор
                description: Аутентификация для пользователя с ролью admin
                value:
                  login: "admin"
                  password: "secretAdminPass"
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Неверный логин или пароль
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Выход пользователя
      description: Деактивирует текущий JWT-токен (добавляет в blacklist на стороне сервера).
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Успешный выход (No Content)
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /operation/all:
    get:
      x-role-required: ORDINARY
      tags:
        - Operations
      summary: Получить список всех операций
      description: Возвращает список всех доступных операций. Доступно для пользователей с ролью ordinary.
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Список операций успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OperationResponse'
        '401':
          description: Не авторизован
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Доступ запрещен (требуется роль ordinary)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /operation/create:
    post:
      x-role-required: ADMIN
      tags:
        - Operations
      summary: Создать новую операцию
      description: Создает новую операцию в системе. Доступно только для администраторов.
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewOperationRequest'
            examples:
              validOperation:
                summary: "Валидная операция"
                value:
                  operationName: "Банковский перевод"
      responses:
        '200':
          description: Операция успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
        '400':
          description: Ошибка валидации (некорректное имя операции)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Не авторизован
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Доступ запрещен (требуется роль ADMIN)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '409':
          description: Операция с таким именем уже существует
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    NewOperationRequest:
      type: object
      required:
        - operationName
      properties:
        operationName:
          type: string
          description: Название операции
          minLength: 1
          maxLength: 100
          example: "Банковский перевод"
      additionalProperties: false
    MediaMethod:
      type: string
      enum: [ sms, email, max, file ]
      description: |
        Способ доставки кода:
        * `sms` - SMS сообщение
        * `email` - Электронная почта
        * `max` - Мессенджер Max
        * `file` - Сохранение в файл
      example: "sms"

    # Общая схема для ID операций и пользователей
    PositiveInt64:
      type: integer
      format: int64
      minimum: 1
      description: "Положительное целое число (ID). Должно быть > 0."
    OtpCode:
      type: string
      pattern: '^[0-9]{4,10}$' # Только цифры, длина от 4 до 10
      example: "0512"
      description: "OTP код. Передается как строка, чтобы сохранить ведущие нули."
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string
        role:
          type: string
          enum: ["admin", "ordinary"]
    RegisterRequest:
      type: object
      required: [login, password, role, email]
      properties:
        login:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 8
          writeOnly: true
          pattern: "^[a-zA-Z0-9!@#$%^&*()]{8,}$"
        role:
          type: string
          enum: ["admin", "ordinary"]
        email:
          type: string
          format: email
          pattern: '^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Za-z]{2,}$'
          example: "user@example.com"
    RegisterResponse:
      type: object
      required: [id, login, role]
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string
        role:
          type: string
          enum: ["admin", "ordinary"]
    LoginRequest:
      type: object
      required: [login, password]
      properties:
        login:
          type: string
          example: "user123"
        password:
          type: string
          format: password
          example: "StrongPass123!"
    LoginResponse:
      type: object
      required: [token, login]
      properties:
        token:
          type: string
        login:
          type: string
        expiresIn:
          type: integer
          description: Срок действия в секундах
          example: 3600
    OtpConfigResponse:
      type: object
      required:
        - lifespan
        - numberOfSymbols
      properties:
        lifespan:
          type: string
          description: ISO-8601 duration (например, PT15M)
          example: "PT15M"
        numberOfSymbols:
          type: integer
          minimum: 4
          description: Количество знаков в OTP
          example: 6
    OtpConfigUpdateRequest:
      type: object
      required:
        - lifespan
        - numberOfSymbols
      properties:
        lifespan:
          type: string
          description: ISO-8601 duration (например, PT15M)
          example: "PT15M"
        numberOfSymbols:
          type: integer
          minimum: 4
          description: Количество знаков в OTP
          example: 6
    ProblemDetail:
      type: object
      required: [status, title]
      properties:
        type:
          type: string
          format: uri
          default: "about:/blank"
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
    OtpGenerateRequest:
      type: object
      required: [operationId, media]
      properties:
        operationId:
          $ref: '#/components/schemas/PositiveInt64'
        media:
          $ref: '#/components/schemas/MediaMethod'
    OtpGenerateResponse:
      type: object
      required:
        - code
        - media
      properties:
        code:
          $ref: '#/components/schemas/OtpCode'
        media:
          $ref: '#/components/schemas/MediaMethod'

    OtpValidateRequest:
      type: object
      required:
        - operationId
        - otpCode
      properties:
        operationId:
          $ref: '#/components/schemas/PositiveInt64'
        otpCode:
          $ref: '#/components/schemas/OtpCode'

    OperationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID операции
          example: 1
        operation_name:
          type: string
          description: Название операции
          example: "Банковский перевод"